requirements.txt
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.0
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
rawpy==0.25.1
scipy==1.13.1
six==1.17.0
tifffile==2024.8.30
tiffile==2018.10.18
zipp==3.23.0
